accessible/tests/mochitest/relations/test_embeds.xhtml

<?xml version="1.0"?>
<?xml-stylesheet href="chrome://global/skin" type="text/css"?>
<?xml-stylesheet href="chrome://mochikit/content/tests/SimpleTest/test.css"
                 type="text/css"?>

<window xmlns="http://www.mozilla.org/keymaster/gatekeeper/there.is.only.xul"
        title="Embeds relation tests">

  <script src="chrome://mochikit/content/tests/SimpleTest/SimpleTest.js" />
  <script src="chrome://mochikit/content/tests/SimpleTest/EventUtils.js"></script>

  <script type="application/javascript"
          src="../common.js"></script>
  <script type="application/javascript"
          src="../role.js"></script>
  <script type="application/javascript"
          src="../states.js"></script>
  <script type="application/javascript"
          src="../events.js"></script>
  <script type="application/javascript"
          src="../relations.js"></script>
  <script type="application/javascript"
          src="../browser.js"></script>

  <script type="application/javascript">
  <![CDATA[
    ////////////////////////////////////////////////////////////////////////////
    // Invokers

    function loadURI(aURI)
    {
      this.invoke = function loadURI_invoke()
      {
        tabBrowser().loadURI(aURI, {
          triggeringPrincipal: Services.scriptSecurityManager.getSystemPrincipal(),
        });
      }

      this.eventSeq = [
        new invokerChecker(EVENT_DOCUMENT_LOAD_COMPLETE, currentTabDocument)
      ];

      this.finalCheck = function loadURI_finalCheck()
      {
        testRelation(browserDocument(), RELATION_EMBEDS,
                     getAccessible(currentTabDocument()));
      }

      this.getID = function loadURI_getID()
      {
        return "load uri " + aURI;
      }
    }

    function loadOneTab(aURI)
    {
      this.invoke = function loadOneTab_invoke()
      {
        tabBrowser().loadOneTab(aURI, {
          referrerURI: null,
          charset: null,
          postData: null,
          inBackground: false,
          triggeringPrincipal: Services.scriptSecurityManager.getSystemPrincipal(),
        });
      }

      this.eventSeq = [
        new invokerChecker(EVENT_DOCUMENT_LOAD_COMPLETE, currentTabDocument)
      ];

      this.finalCheck = function loadURI_finalCheck()
      {
        testRelation(browserDocument(), RELATION_EMBEDS,
                     getAccessible(currentTabDocument()));
      }

      this.getID = function loadOneTab_getID()
      {
        return "load uri '" + aURI + "' in new tab";
      }
    }

    ////////////////////////////////////////////////////////////////////////////
    // Testing

    //gA11yEventDumpToConsole = true; // debug

    var gQueue = null;
    function doTests()
    {
      testRelation(browserDocument(), RELATION_EMBEDS,
                   getAccessible(currentTabDocument()));

      enableLogging("docload");
      gQueue = new eventQueue();

      gQueue.push(new loadURI("about:robots"));
      gQueue.push(new loadOneTab("about:mozilla"));

      gQueue.onFinish = function()
      {
        disableLogging();
        closeBrowserWindow();
      }
      gQueue.invoke();
    }

    SimpleTest.waitForExplicitFinish();
    openBrowserWindow(doTests, "about:license");
  ]]>
  </script>

  <vbox flex="1" style="overflow: auto;">
  <body xmlns="http://www.w3.org/1999/xhtml">
    <a target="_blank"
       href="https://bugzilla.mozilla.org/show_bug.cgi?id=707654"
       title="Embeds relation on root accessible can return not content document">
      Mozilla Bug 707654
    </a>
    <p id="display"></p>
    <div id="content" style="display: none">
    </div>
    <pre id="test">
    </pre>
  </body>
  </vbox>
</window>
